fix(db): widen observations.captured_at_ms to BigInteger
Epoch milliseconds (Date.now() ~1.7e12) overflow a 32-bit Integer, so inserting
any real frida/mitmproxy observation threw and rolled back the entire dynamic
transaction -- silently losing all runtime evidence (same class as the JSONB
mutation bug). Widen to BigInteger (migration 0006).

+++ parallax/parallax/core/models/observation.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, Text, func
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, String, Text, func
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -41,7 +41,9 @@ class Observation(Base):
     exception: Mapped[str | None] = mapped_column(Text, nullable=True)
 
     # Timing
-    captured_at_ms: Mapped[int] = mapped_column(Integer, nullable=False)
+    # Epoch milliseconds — must be BigInteger; a 32-bit Integer overflows
+    # (Date.now() ~1.7e12) and the insert rolls back the whole dynamic txn.
+    captured_at_ms: Mapped[int] = mapped_column(BigInteger, nullable=False)
     session_id: Mapped[str | None] = mapped_column(String(128), nullable=True)
 
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
